+++ package.json
@@ -17,7 +17,6 @@
   },
   "keywords": [],
   "author": "",
-  "type": "module",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/papaAndrew/agapitov-andrew-otus/issues"
